Problems regarding the last commit.
It said:
" [...] Wrote implementation for most CharSet functions in 'CharSet.c'. ... "
" [...] There is still no implementation for 'pf_CharSet_sort'. "

'CharSet_make' is actually supposed to call 'pf_CharSet_sort' to randomize it's contents.
Now it just initializes it will the letter of the alphabet in order.

Also, the random IDs of the nodes seem to not work. They're zeros and ones.

Issues were raised for each of these problems.